Move ProgressHandler to same file with createProgressBar
